requirements.txt
numpy>=2.2.2
matplotlib.pyplot>=3.10.0
